Merge pull request #49579 from deads2k/server-34-default
Automatic merge from submit-queue (batch tested with PRs 55798, 49579, 54862, 55188, 51990). If you want to cherry-pick this change to another branch, please follow the instructions <a href="https://github.com/kubernetes/community/blob/master/contributors/devel/cherry-picks.md">here</a>.

add back defaulting for parameter decoding

At the beginning of 1.7, we removed the last "conversion causes defaulting".  This broke the "default to true" behavior for exec and attach options, but we didn't notice.  This removes the broken defaulter (you can default a non-point bool to true on an object) and adds back defaulting to parameter codecs.

@k8s-mirror-api-machinery-misc @lavalamp @smarterclayton

Kubernetes-commit: 3c51f539609ac8bc0c74fc859570e44a7f07894e